test(c): split test_mode.py integration -> integration/test_mode.py (#40)
Moved 11 tests that invoke pipeline.run_all via the local _run_pipeline
helpers — they exercise mode flag semantics (dev/prod/test) across
pipeline + config + state + destinations. Remaining 8 config-parsing
and env-var tests stay in tests/test_mode.py for Wave D.

Helpers _run_pipeline and _run_with_transforms moved along with the
tests that use them.
